polarization-analysis/requirements.txt
google-generativeai
google-api-python-client

stable-ts
pyannote.audio==3.1.1
torch
torchaudio

pandas
tqdm